Convert math structures back to float, see issue #2653. This gives us a small but noticeable performance boost of about 10-20% and smaller memory footprint, and it leaves the door open to using SIMD instructions more effectively sometime in the future.

+++ include/icommandsystem.h
@@ -76,8 +76,8 @@ public:
 		_strValue(doubleToStr(d)),
 		_doubleValue(d),
 		_intValue(static_cast<int>(d)),
-		_vector3Value(d,d,d),
-		_vector2Value(d,d),
+		_vector3Value(static_cast<float>(d),static_cast<float>(d),static_cast<float>(d)),
+		_vector2Value(static_cast<float>(d),static_cast<float>(d)),
 		_type(ARGTYPE_DOUBLE)
 	{
 		// Enable INT flag if double value is rounded
@@ -91,8 +91,8 @@ public:
 		_strValue(intToStr(i)),
 		_doubleValue(static_cast<double>(i)),
 		_intValue(i),
-		_vector3Value(i,i,i),
-		_vector2Value(i,i),
+		_vector3Value(static_cast<float>(i),static_cast<float>(i),static_cast<float>(i)),
+		_vector2Value(static_cast<float>(i),static_cast<float>(i)),
 		_type(ARGTYPE_INT|ARGTYPE_DOUBLE) // INT can be used as DOUBLE too
 	{}
