+++ libs/scenelib.h
@@ -27,7 +27,7 @@ class SelectionTest;
 class VolumeTest;
 template<typename Element>
 class BasicVector4;
-typedef BasicVector4<double> Vector4;
+typedef BasicVector4<float> Vector4;
 class Matrix4;
 typedef Vector4 Quaternion;
 class AABB;
